Use another method to get ImageView in NavBar
Remove calling “getXXXButton” which may be removed in some ROM’s
(e.g. CyanogenMod)
& also because we catch the NoSuchMethodError and use
“findViewWithTag”, it may be slow

Instead, in this method, we do a loop to search for ImageView’s and
caches them in an ArrayList.
This will increase compatibility and also tint other
icons that we have not covered
(eg. When you add “Search” button in custom ROM’s, it was not tinted,
but this method should tint it too)